fix gs bug
